% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive_metrics.R
\name{calculate_global_audit_pro}
\alias{calculate_global_audit_pro}
\title{Updated Global Audit with Specialty Grouping}
\usage{
calculate_global_audit_pro(df)
}
\description{
Updated Global Audit with Specialty Grouping
}
