% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_test.R
\name{calculate_MAFI}
\alias{calculate_MAFI}
\title{Calculate MAFI (Meta-Analysis Fragility Index)}
\usage{
calculate_MAFI(
  yi,
  vi,
  measure = "SMD",
  clinical_threshold = NULL,
  alpha = 0.05
)
}
\description{
Calculate MAFI (Meta-Analysis Fragility Index)
}
