% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohorts.R
\name{get_cohort}
\alias{get_cohort}
\title{Select a coherent cohort for analysis}
\usage{
get_cohort(df, outcome, measure, drop_na = TRUE)
}
\arguments{
\item{df}{effects table from read_mlm_effects()}

\item{outcome}{one of c("DICH","CONT","GENSUM")}

\item{measure}{character (e.g., "logOR","MD","GIV","GEN_CI")}

\item{drop_na}{logical: drop rows with NA TE or seTE}
}
\value{
filtered data.frame
}
\description{
Select a coherent cohort for analysis
}
