% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_cohort.R
\name{plot_cohort_gg}
\alias{plot_cohort_gg}
\title{Plot a cohort with ggplot2 (forest-style)}
\usage{
plot_cohort_gg(df, outcome, measure, top_n = 50, title = NULL)
}
\arguments{
\item{df}{effects table (from read_mlm_effects())}

\item{outcome}{one of c("DICH","CONT","GENSUM")}

\item{measure}{e.g., "logOR","MD","SMD","GIV","GEN_CI"}

\item{top_n}{integer: only show the top-N most precise (smallest seTE)}

\item{title}{plot title}
}
\description{
Plot a cohort with ggplot2 (forest-style)
}
