% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_cohort.R
\name{plot_cohort_quick}
\alias{plot_cohort_quick}
\title{Plot a cohort quickly (base R)}
\usage{
plot_cohort_quick(df, outcome, measure, main = NULL)
}
\arguments{
\item{df}{effects table (from read_mlm_effects())}

\item{outcome}{one of c("DICH","CONT","GENSUM")}

\item{measure}{e.g., "logOR","MD","SMD","GIV","GEN_CI"}

\item{main}{plot title}
}
\description{
Plot a cohort quickly (base R)
}
