% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_gg.R
\name{plot_effects_gg}
\alias{plot_effects_gg}
\title{Forest-style plot with ggplot2}
\usage{
plot_effects_gg(df, top_n = 50, title = "Effects (TE ± 1.96*SE)")
}
\arguments{
\item{df}{data.frame with columns TE, seTE and optional id}

\item{top_n}{integer: show top N by precision (default: 50)}

\item{title}{character: plot title}
}
\description{
Forest-style plot with ggplot2
}
