% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_effects_quick}
\alias{plot_effects_quick}
\title{Quick effects plot (base R)}
\usage{
plot_effects_quick(df, main = "Effects (TE ± 1.96*SE)")
}
\arguments{
\item{df}{data.frame with columns TE, seTE}

\item{main}{plot title}
}
\description{
Quick effects plot (base R)
}
