% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{read_mlm_effects}
\alias{read_mlm_effects}
\title{Read the multilevel effects table}
\usage{
read_mlm_effects()
}
\value{
data.frame with columns review_id, analysis_id, study_id, TE, seTE, etc.
}
\description{
Read the multilevel effects table
}
