% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.R
\name{simulate_pathway}
\alias{simulate_pathway}
\title{Simulate a clinical pathway from diagnosis to outcome.}
\usage{
simulate_pathway(
  prevalence,
  sensitivity,
  specificity,
  rr,
  harm_rate = 0,
  n = 1000
)
}
\description{
Simulate a clinical pathway from diagnosis to outcome.
}
